requirements.txt
paho-mqtt
pillow
requests